move state to issues
